@stoksc 2023-02-09
i would not label this 'user facing api change' since it should not change the api presented to users, only its implementation.